Updated logic to avoid EOFException looping
This IOException added due to getting Error EOFException and in this
next and last variables were not initialized due to error. And it was
going into infinite loop. As this process in hang state, a subsequent
process was not kicking off and no more data were receiving. Also,
Error log was getting full as with the loop it was logging same errors
in each loop.

In this setting next to 'null' and last to 'true' to come out of the
loop gracefully.